+++ odoo/local-src/specific_sale/models/__init__.py
@@ -7,3 +7,4 @@ from . import res_partner
 from . import sale
 from . import sale_report
 from . import stock_production_lot
+from . import product_category