+++ odoo/local-src/specific_sale/tests/__init__.py
@@ -1,4 +1,4 @@
 # -*- coding: utf-8 -*-
 
-# TODO: See why this test fail
-# from . import test_sale_order_line_qty_unavailable
+from . import test_sale_order_line_qty_unavailable
+from . import test_sale_exceptions